Updated what we use for vehicle names
We continue to use the page title initially - the vehicle page might not
have an info box so this is ok. However when we confirm that the page
has an inbox we use the vehicle name from the title section as this
USUALLY matches the link title on the main page - there are exceptions
to this so the next commit will be changing to use the link titles
instead.