Beginning stage one of refactor re #3
Created ILogger which will handle file interactions.
Created IConsoleManager which handles console interactions.
Updated all direct references to Console to point to an instance of
ConsoleManager in preparation for when the rest of the Main() is
refactored and ConsoleManager is provided via constructor injection.
Added application version to the top of the the console when it begins.

Refs #3